LPS-75609 Move everything to internal
